now vmc compiles with new dilprg class
