+++ vme/src/vmc/dilshare.cpp
@@ -1,0 +1,11 @@
+/*
+ $Author: All $
+ $RCSfile: structs.cpp,v $
+ $Date: 2004/05/16 04:34:34 $
+ $Revision: 2.4 $
+ */
+/*this file is only here because make file is a pain in the butt.  Once the system is switched
+to use Jamfiles this file can be removed.
+*/
+
+#include "../dilshare.cpp"
